backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector
sentence-transformers
beautifulsoup4
requests
numpy 